[PATCH: gcc3.3-signed-fix.dif]
    6c7c9be6501838d74508f49dcb80b4bb02653e6c
- ver.0.9.0rc7. we took a step forward...
    270aadfc6ffd7c60f1c47dfceff6788f007be705
Replaced n with nothing for acinclude
    7e927318a2b38cd46ba8c60f04c239ae6a8b5c52
Next fix for makefiles
    f8b5b36044e135e4a0fbdab952c4835489c40eb8
Added DT0398 codec???
    e2966f7a234e09eca1471afb173f26c6d5357a2c
Replaced snd_dma_residue() with snd_dma_pointer().
    8e8c43e573c424fc9b34ae20e189081ef3def60c
Fixed typo
    a9037c85f6e4c628ec1b45420908423ca8f67a76
Fixed sequencer macros (not build unwanted modules)
    de8603b3a141d655f62e1c3f983bf55636d7234a
Anders Torger <torger@ludd.luth.se> fixed bug introduced in previous patch - analog input now works again
    5c0cd1bd86c7d334f10e309de6c947d7dedf1f1e
- fixed multi-channel playback.  now channel B is used for that. - check the devices for multi-channel playback.  both channels are necessary. - fixed the configuration of period size.
    03fc1200ef2726029d47a4a1afc6c5d5bb767072

+++ sound/core/seq/seq_timer.c
@@ -180,7 +180,7 @@ int snd_seq_timer_set_tempo(seq_timer_t * tmr, int tempo)
 	if (tempo <= 0)
 		return -EINVAL;
 	spin_lock_irqsave(&tmr->lock, flags);
-	if (tempo != tmr->tempo) {
+	if ((unsigned int)tempo != tmr->tempo) {
 		tmr->tempo = tempo;
 		snd_seq_timer_set_tick_resolution(&tmr->tick, tmr->tempo, tmr->ppq, 1);
 	}
